Remove tests that fail with Swift 3 expression type checking support removed.
Most of these are tested with 4.x as well, and remain. I've wholesale
deleted the ones that are only tested with 3.x, but could revert those
parts of XFAIL them instead if there is a reason to do so. My thoughts
are that it's easy to replace these when we get hashes to text 4.x with.