Temporarily remove Kickstarter-ReactiveExtensions and RxDataSources.
We do not currently have Swift 4.x hashes for these.

Kickstarter is being worked on in
https://github.com/apple/swift-source-compat-suite/pull/211, but that
is not ready yet.

I tried to update RxDataSources myself, but ran into issues because
updating the hash means also updating other parts of the build, and I
was unable to figure out how to do that.

After I merge https://github.com/apple/swift/pull/17691, I will revert
this commit and add XFAILs for these 3.x configurations, including
links to JIRAs so that we don't lose track of the fact that we still
need to update these for 4.x.