Remove hack that’s causing cperf failures
Previously, because we chdir()ed before parsing arguments, the --output argument would be relative to the swift-source-compat-suite directory, which necessitated a hack to write the comment file to the CI-expected path. Commit 6952162 accidentally fixed that bug. Unfortutnately, it also broke the hack, which caused all compiler performance tests to fail. This commit deletes the hack, which should restore correct behavior.