[run_sk_stress_test] List all failures at the end of the run
Suppresses inline progress output and instead dumps the failures at the end of
the run. This is to work around the CI job failing in SwiftPM packages due to
the extra inline output sk-swiftc-wrapper produces on top of the wrapped
compiler output not being parseable when SwiftPM builds pass
`-parseable-output`.