Add support for xfailing based on build configuration (release vs debug)
We have a project failing only in release configuration at the moment with no
existing way to xfail only release.

This change also reworks how xfail matching works to be more expressive, since
it could previously only match either branches or platforms for a specific
compatibility version, not both. See the README.md changes for details.

Migrated projects.json to the new structure (projects-incremental.json has no
xfails currently) and updated the starscream xfail to be release-only, but
left support for the old structure to allow time to update the json in
downstream repos.