[Stress Tester] Report XFails whose file was not tested in the current run
If the file covered by an XFail is not tested in the current run, it appears neither in the expected failures, nor in the unexpected failures section.

List these XFails their own section. I deliberately chose not to fail the stress tester if we have one of these XFails because
a) files also end up here if their project isn’t run in the current stress tester run (might be interesting if we include more projects in the weekly run again)
b) the XFail is still valid but it’s shadowed by a different failure which causes the stress tester to not reach the XFailed file

If neither of these reasons applies, we can safely remove an XFail