Manage the project XFails under a different job type for the stress tester
Currently we’re in a state again in which some project succeed to build in the stress tester CI job but fail to build in the source compatibility suite CI job (rdar://82333521). We are thus mistakingly receiving UPASSes in the stress tester CI job.

To work around this, introduce the notion of job types to the source compatibility runner and its associated XFails. By default everything runs as the "source-compat" job, only the stress tester overrides the job to "stress-tester". That way we can selectively XFail jobs only for the stress tester in `projects.json`.

Since the stress tester now controls its own project build XFails, a project build failure will also cause a stress tester failure.

rdar://62080167