Add https://github.com/abertelrud/swift-protobuf-plugin-example.git to the source compatibility suite.
This is a test of SwiftPM functionality that is new in SwiftPM 5.6 (thus the XFAIL for `release/5.5`).

This is a sample package that uses a modified version of swift-protobuf to which a SwiftPM build tool plugin has been added.  The sample verifies that `.proto` files get processed properly.

This is currently only Darwin but a future PR will extend it to Linux.

rdar://84315022