Support using `sk-swiftc-wrapper` that is not installed in the toolchain
Previously, when running the stress tester, we changed the `swiftc` propertie in `project_future.py` to `sk-swiftc-wrapper`. But this only worked if `sk-swiftc-wrapper` is installed into the toolchain because it tries to find `swift` next to it.

To support using an `sk-swiftc-wrapper` that’s not installed into the toolchain, keep the `swiftc` property pointing to the actual `swiftc` executable in the toolchain and add a new option with which we override the `SWIFT_EXEC` that is actually used to build the projects to `sk-swiftc-wrapper`.