Remove Evergreen
The Git URL of Evergreen no longer exists, it redirects to NetNewsWire. Since the project has also been renamed and there is no longer an Xcode workspace, I’m not sure if they are the same projects. If NetNewsWire should be part of the source compat suite, it should be added agai.