Use a preset for the build-script invocation used by `run_sk_stress_test`
The hard-coded `build-script` invocation is gross and also didn’t build the new driver, which lead to issues. Use a preset based on the source compat suite presets instead.

Effectively, this will add the following arguments to the `build-script` invocation. `--llvm-install-components` and `--swift-install-components` just show the components the newly added components.

```
--cross-compile-hosts=macosx-arm64
--install-llvm \
--install-swift-driver \
--llvm-install-components=clang;clang-resource-headers;compiler-rt;clang-features-file
--swift-driver \
--swift-install-components=back-deployment;libexec
```